refactor: Regroup ShareModal into its own folder into components
