+++ packages/cozy-sharing/src/index.jsx
@@ -14,7 +14,7 @@ export { SharingOwnerAvatar } from './SharingOwnerAvatar'
 export { SharedRecipients } from './SharedRecipients'
 export { SharedRecipientsList } from './SharedRecipientsList'
 export { ShareButton } from './ShareButton'
-export { ShareModal } from './ShareModal'
+export { ShareModal } from './components/ShareModal/ShareModal'
 export { RefreshableSharings } from './RefreshableSharings'
 export { useFetchDocumentPath } from './components/useFetchDocumentPath'
 export { CozyPassFingerprintDialogContent } from './components/CozyPassFingerprintDialogContent'
